Done Task Dec 6, 2021. Removed a lot of let calls because I don't know how to use it properly. Almost Finished with Order and Order Detail Fragment.

+++ app/src/main/java/com/teampym/onlineclothingshopapplication/data/room/Product.kt
@@ -44,7 +44,7 @@ data class Product(
     var flag: String = ""
 
     @get:Exclude
-    val avgRate: BigDecimal get() = (totalRate / numberOfReviews).toBigDecimal()
+    val avgRate: BigDecimal get() = (totalRate / numberOfReviews.toDouble()).toBigDecimal()
 
     @get:Exclude
     @Ignore
